refactor: Extract mock for cozy-client query
