+++ src/drive/web/modules/drive/Toolbar/components/MoreMenu.spec.jsx
@@ -3,12 +3,10 @@ import { mount } from 'enzyme'
 import flag from 'cozy-flags'
 import { act } from 'react-dom/test-utils'
 import { ActionMenuItem } from 'cozy-ui/transpiled/react/ActionMenu'
-import { setupFolderContent } from 'test/setup'
+import { setupFolderContent, mockCozyClientRequestQuery } from 'test/setup'
 import { downloadFiles } from 'drive/web/modules/actions/utils'
 import MoreMenu from 'drive/web/modules/drive/Toolbar/components/MoreMenu'
 import AppLike from 'test/components/AppLike'
-import { generateFile } from 'test/generate'
-import CozyClient from 'cozy-client'
 import { MoreButton } from 'components/Button'
 
 jest.mock('drive/web/modules/actions/utils', () => ({
@@ -17,18 +15,7 @@ jest.mock('drive/web/modules/actions/utils', () => ({
 
 const sleep = duration => new Promise(resolve => setTimeout(resolve, duration))
 
-beforeEach(() => {
-  const files = Array(10)
-    .fill(null)
-    .map((x, i) => generateFile({ i }))
-  const directories = Array(3)
-    .fill(null)
-    .map((x, i) => generateFile({ i, type: 'directory' }))
-  const fileAndDirs = directories.concat(files)
-  jest.spyOn(CozyClient.prototype, 'requestQuery').mockResolvedValue({
-    data: fileAndDirs
-  })
-})
+mockCozyClientRequestQuery()
 
 describe('MoreMenu v2', () => {
   beforeEach(() => {
